requirements.txt
# Core ML/AI Frameworks
torch>=2.0.0
transformers>=4.35.0
peft>=0.7.0
sentence-transformers>=2.2.2

# Database
psycopg2-binary>=2.9.9
pgvector>=0.2.3

# Data Processing
numpy>=1.24.0

# Configuration and Utilities
pyyaml>=6.0.1
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Monitoring and Logging
loguru>=0.7.2

# Testing
pytest>=7.4.3
pytest-cov>=4.1.0
pytest-asyncio>=0.21.1

# Code Quality
black>=23.11.0
flake8>=6.1.0
mypy>=1.7.0
